requirements.txt
Flask==2.2.2
Werkzeug==2.3.7
pyyaml
transformers == 4.41.2
sentencepiece ==0.2.0
sentence-transformers == 3.0.0

faiss-cpu
pandas





